Fix bundled types
