--- generators/base/templates/@types/index.d.ts
@@ -1,10 +1,0 @@
-interface ReducerAction {
-  type: string;
-}
-
-type RootStackParamList = {
-  Main: undefined;
-  Styleguide: undefined;
-};
-
-type Environment = "development" | "staging" | "edge" | "production" | "live";
