--- generators/base/templates/@types/styled.d.ts
@@ -1,14 +1,0 @@
-import "styled-components";
-
-declare module "styled-components" {
-  export interface DefaultTheme {
-    spacing: number;
-    fonts: {
-      family: string;
-    };
-    colors: {
-      white: string;
-      grey: string;
-    };
-  }
-}
