+++ generators/base/templates/App/Router.tsx
@@ -3,6 +3,11 @@ import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
 import Screens from "<%= name %>/App/Screens";
 
+export type RootStackParamList = {
+  Main: undefined;
+  Styleguide: undefined;
+};
+
 const Stack = createStackNavigator<RootStackParamList>();
 
 function RootStack() {
