+++ generators/base/templates/App/Screens/Styleguide.tsx
@@ -5,6 +5,7 @@ import styled from "styled-components/native";
 import Button from "<%= name %>/App/Components/Button";
 import Layout from "<%= name %>/App/Components/Layout";
 import Text from "<%= name %>/App/Components/Text";
+import { RootStackParamList } from "<%= name %>/App/Router";
 
 type StyleguideScreenNavigationProp = StackNavigationProp<
   RootStackParamList,
