+++ generators/base/templates/App/Theme.ts
@@ -1,4 +1,19 @@
-export default {
+import { DefaultTheme } from "styled-components/native";
+
+declare module "styled-components" {
+  export interface DefaultTheme {
+    spacing: number;
+    fonts: {
+      family: string;
+    };
+    colors: {
+      white: string;
+      grey: string;
+    };
+  }
+}
+
+const theme: DefaultTheme = {
   spacing: 16,
   fonts: {
     family: "System",
@@ -8,3 +23,5 @@ export default {
     grey: "grey",
   },
 };
+
+export default theme;
